add layer and e2e test for easy_slack_notify

+++ tests/e2e/test_e2e.py
@@ -1,102 +1,27 @@
 import json
-import os
 
-import boto3
 import pytest
-from botocore.exceptions import ClientError
 
 
-@pytest.fixture(scope='module')
-def stack_name():
-    return os.environ['STACK_NAME']
-
-
-@pytest.fixture(scope='module')
-def cfn():
-    return boto3.client('cloudformation')
-
-
-@pytest.fixture(scope='module', autouse=True)
-def prepare_stack(stack_name, cfn):
-    template = None
-    with open('.sam/template.yml') as f:
-        template = f.read()
-
-    deploy = cfn.update_stack
-    waiter = cfn.get_waiter('stack_update_complete')
-
-    try:
-        cfn.describe_stacks(StackName=stack_name)
-    except ClientError as e:
-        if e.response['Error']['Message'] == f'Stack with id {stack_name} does not exist':
-            deploy = cfn.create_stack
-            waiter = cfn.get_waiter('stack_create_complete')
-        else:
-            raise
-
-    options = {
-        'StackName': stack_name,
-        'TemplateBody': template,
-        'Capabilities': [
-            'CAPABILITY_IAM',
-            'CAPABILITY_AUTO_EXPAND'
-        ]
-    }
-    deploy(**options)
-    waiter.wait(StackName=stack_name)
-
-    yield
-
-    cfn.delete_stack(StackName=stack_name)
-    cfn.get_waiter('stack_delete_complete').wait(StackName=stack_name)
-
-
-@pytest.fixture(scope='module')
-def sqs():
-    return boto3.client('sqs')
-
-
-@pytest.fixture(scope='module')
-def sns():
-    return boto3.client('sns')
-
-
-@pytest.fixture(scope='module')
-def stack_info(stack_name, cfn):
-    resp = cfn.describe_stacks(StackName=stack_name)
-    return resp['Stacks'][0]
-
-
-@pytest.fixture(scope='module')
-def stack_outputs(stack_info):
-    result = {}
-    for item in stack_info['Outputs']:
-        result[item['OutputKey']] = item['OutputValue']
-    return result
-
-
-@pytest.fixture(scope='module')
-def sns_topic_arn(stack_outputs):
-    return stack_outputs['SNSTopicArn']
-
-
-@pytest.fixture(scope='module')
-def sqs_url(stack_outputs):
-    return stack_outputs['SQSUrl']
-
-
-@pytest.fixture(scope='module')
-def specific_url(stack_outputs):
-    return stack_outputs['SpecificUrl']
-
-
-def get_sqs_message(sqs, sqs_url):
+def get_sqs_message(sqs, sqs_url, will_delete=True):
     resp = sqs.receive_message(QueueUrl=sqs_url, MaxNumberOfMessages=1)
     record = resp['Messages'][0]
     receipt_handle = record['ReceiptHandle']
     body = record['Body']
-    sqs.delete_message(QueueUrl=sqs_url, ReceiptHandle=receipt_handle)
+    if will_delete:
+        sqs.delete_message(QueueUrl=sqs_url, ReceiptHandle=receipt_handle)
     return json.loads(body)
+
+
+def lambda_invoke(lambda_client, function_name, event):
+    options = {
+        'FunctionName': function_name,
+        'InvocationType': 'RequestResponse',
+        'Payload': json.dumps(event).encode('utf-8')
+    }
+    resp = lambda_client.invoke(**options)
+    if resp['StatusCode'] != 200:
+        raise Exception('Lambda failed')
 
 
 list_expected = [
@@ -129,6 +54,53 @@ list_expected = [
     )
 ]
 
+test_data = {
+    'layer': {
+        'easy_slack_notify': [
+            (
+                {
+                    'message': 'sinofseven'
+                },
+                {
+                    'text': 'sinofseven'
+                }
+            ),
+            (
+                {
+                    'message': 'あなたは夜明けに微笑んで',
+                    'channel': '#first'
+                },
+                {
+                    'text': 'あなたは夜明けに微笑んで',
+                    'channel': '#first'
+                }
+            ),
+            (
+                {
+                    'message': '奏でる少女の道行きは',
+                    'username': 'second'
+                },
+                {
+                    'text': '奏でる少女の道行きは',
+                    'username': 'second'
+                }
+            ),
+            (
+                {
+                    'message': 'アマデウスの詩、謳え敗者の王',
+                    'channel': '#third',
+                    'username': 'third'
+                },
+                {
+                    'text': 'アマデウスの詩、謳え敗者の王',
+                    'channel': '#third',
+                    'username': 'third'
+                }
+            )
+        ]
+    }
+}
+
 
 class TestDefaultUrl(object):
     @pytest.mark.parametrize('expected', list_expected)
@@ -149,4 +121,214 @@ class TestSpecificUrl(object):
         body = get_sqs_message(sqs, sqs_url)
 
         assert body['Subject'] == 'specific'
+        assert json.loads(body['Message']) == expected
+
+
+class TestLayerPython36(object):
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_default(self, data, expected, lambda_client, name_lambda_python36, sqs, sqs_url):
+        event = {
+            'case': 'test_easy_slack_notify_default',
+            'data': data
+        }
+        lambda_invoke(lambda_client, name_lambda_python36, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_specific_url(self, data, expected, lambda_client, name_lambda_python36,
+                                            sqs, sqs_url, specific_url):
+        event = {
+            'case': 'test_easy_slack_notify_specific_url',
+            'data': data,
+            'url': specific_url
+        }
+        lambda_invoke(lambda_client, name_lambda_python36, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'specific'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_parameter_name(self, data, expected, lambda_client, name_lambda_python36,
+                                                  sqs, sqs_url, another_ssm_parameter_name):
+        event = {
+            'case': 'test_easy_slack_notify_set_parameter_name',
+            'data': data,
+            'ssm_parameter_name': another_ssm_parameter_name
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python36, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_topic_arn(self, data, expected, lambda_client, name_lambda_python36,
+                                             sqs, sqs_url, sns_topic_arn):
+        event = {
+            'case': 'test_easy_slack_notify_set_topic_arn',
+            'data': data,
+            'topic_arn': sns_topic_arn
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python36, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+
+class TestLayerPython37(object):
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_default(self, data, expected, lambda_client, name_lambda_python37, sqs, sqs_url):
+        event = {
+            'case': 'test_easy_slack_notify_default',
+            'data': data
+        }
+        lambda_invoke(lambda_client, name_lambda_python37, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_specific_url(self, data, expected, lambda_client, name_lambda_python37,
+                                            sqs, sqs_url, specific_url):
+        event = {
+            'case': 'test_easy_slack_notify_specific_url',
+            'data': data,
+            'url': specific_url
+        }
+        lambda_invoke(lambda_client, name_lambda_python37, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'specific'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_parameter_name(self, data, expected, lambda_client, name_lambda_python37,
+                                                  sqs, sqs_url, another_ssm_parameter_name):
+        event = {
+            'case': 'test_easy_slack_notify_set_parameter_name',
+            'data': data,
+            'ssm_parameter_name': another_ssm_parameter_name
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python37, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_topic_arn(self, data, expected, lambda_client, name_lambda_python37,
+                                             sqs, sqs_url, sns_topic_arn):
+        event = {
+            'case': 'test_easy_slack_notify_set_topic_arn',
+            'data': data,
+            'topic_arn': sns_topic_arn
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python37, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+
+class TestLayerPython27(object):
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_default(self, data, expected, lambda_client, name_lambda_python27, sqs, sqs_url):
+        event = {
+            'case': 'test_easy_slack_notify_default',
+            'data': data
+        }
+        lambda_invoke(lambda_client, name_lambda_python27, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_specific_url(self, data, expected, lambda_client, name_lambda_python27,
+                                            sqs, sqs_url, specific_url):
+        event = {
+            'case': 'test_easy_slack_notify_specific_url',
+            'data': data,
+            'url': specific_url
+        }
+        lambda_invoke(lambda_client, name_lambda_python27, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'specific'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_parameter_name(self, data, expected, lambda_client, name_lambda_python27,
+                                                  sqs, sqs_url, another_ssm_parameter_name):
+        event = {
+            'case': 'test_easy_slack_notify_set_parameter_name',
+            'data': data,
+            'ssm_parameter_name': another_ssm_parameter_name
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python27, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
+        assert json.loads(body['Message']) == expected
+
+    @pytest.mark.parametrize(
+        'data, expected', test_data['layer']['easy_slack_notify']
+    )
+    def test_easy_slack_notify_set_topic_arn(self, data, expected, lambda_client, name_lambda_python27,
+                                             sqs, sqs_url, sns_topic_arn):
+        event = {
+            'case': 'test_easy_slack_notify_set_topic_arn',
+            'data': data,
+            'topic_arn': sns_topic_arn
+        }
+
+        lambda_invoke(lambda_client, name_lambda_python27, event)
+
+        body = get_sqs_message(sqs, sqs_url)
+
+        assert body['Subject'] == 'default'
         assert json.loads(body['Message']) == expected
